refactor: 코스 DTO 관련 클래스들을 하나로 통합
