+++ android/app/src/main/java/io/coursepick/coursepick/data/CourseService.kt
@@ -8,5 +8,5 @@ interface CourseService {
     suspend fun courses(
         @Query("lat") latitude: String,
         @Query("lng") longitude: String,
-    ): List<CoursesResponseItem>
+    ): List<CourseResponse>
 }
