--- android/app/src/main/java/io/coursepick/coursepick/data/Geometry.kt
@@ -1,9 +1,0 @@
-package io.coursepick.coursepick.data
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Geometry(
-    val coordinates: List<List<Double?>?>?,
-    val type: String?,
-)
